[add_serial_terminal] moved hash map and commands to their own file
