--- kernel/include/kernel/serialterm/serial_terminal.h
@@ -1,6 +1,0 @@
-#ifndef _KERNEL_SERIAL_CONSOLE_H
-#define _KERNEL_SERIAL_CONSOLE_H
-
-int serial_terminal(const char *prompt);
-
-#endif
